fix: verification and optimization pass
- Fix SHA-256 round constant k[56] typo (0x748f82ae -> 0x748f82ee)
- Fix SHA-256 finalization: proper padding, big-endian bit count
- Fix SHA-256 block word loading: byte-by-byte to avoid alignment UB
- Fix log.c buffer overflow in CR/LF append (off-by-one)
- Fix log.c division by zero guard in itoa_str base validation
- Fix 64-bit host build: REG32 casts through uintptr_t (8 files)
- Fix ARM assembly guards (#ifdef __arm__) for host compilation
- Fix ARM inline asm: stmia -> individual str, mrc operand syntax
- Fix linker: page table moved to DRAM section, heap 16K->4K, stack 32K->16K
- Fix IRQ table: MAX_IRQS 1024->256 to fit SRAM budget
- Add bare-metal memcpy/memset/memcmp/strlen in string.c
- Add include/utils/string.h header
- Add CMAKE_TRY_COMPILE_TARGET_TYPE for bare-metal toolchain
- Add -lgcc for ARM division builtins
- Expand test mock layer with comprehensive HAL/driver stubs
- All 6 host tests pass, ARM cross-compile links clean (11K text)

+++ src/core/loader.c
@@ -4,6 +4,7 @@
  */
 #include "core/loader.h"
 #include "driver/flash.h"
+#include "driver/crc.h"
 #include "errno.h"
 #include "config.h"
 #include "utils/string.h"
@@ -53,13 +54,13 @@ int core_load_image(uint32_t flash_addr, image_header_t *header)
     uint32_t payload_size = raw_header.size;
     uint32_t data_offset = flash_addr + sizeof(raw_image_header_t);
 
-    ret = driver_flash_read(data_offset, (uint8_t *)header->load_addr, payload_size);
+    ret = driver_flash_read(data_offset, (uint8_t *)(uintptr_t)header->load_addr, payload_size);
     if (ret < 0) {
         return BOOT_ERR_IMAGE_LOAD;
     }
 
     /* Verify CRC */
-    uint32_t crc = driver_crc32((uint8_t *)header->load_addr, payload_size, 0);
+    uint32_t crc = driver_crc32((uint8_t *)(uintptr_t)header->load_addr, payload_size, 0);
     if (crc != raw_header.crc32) {
         return BOOT_ERR_IMAGE_CORRUPT;
     }
@@ -96,7 +97,7 @@ int core_load_dtb(uint32_t dtb_addr, uint32_t load_addr)
     }
 
     /* Read entire DTB */
-    ret = driver_flash_read(dtb_addr, (uint8_t *)load_addr, size);
+    ret = driver_flash_read(dtb_addr, (uint8_t *)(uintptr_t)load_addr, size);
     if (ret < 0) {
         return BOOT_ERR_IMAGE_LOAD;
     }
